// frontend/src/api/generated/zod/generationDefinitionSchema.zod.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Drawing API
 * API to manage drawings and their elements
 * OpenAPI spec version: 1.0.0
 */
import * as zod from 'zod';

export const generationDefinitionSchemaStepsByIdVisibilityDefault = `visible`;
export const generationDefinitionSchemaStepsByIdGroupModeDefault = `none`;
export const GenerationDefinitionSchema = zod.object({
  "rootStepId": zod.string(),
  "stepsById": zod.record(zod.string(), zod.object({
  "id": zod.string().optional(),
  "parentId": zod.string().nullish(),
  "childIds": zod.array(zod.string()).optional(),
  "relationship": zod.string().nullish(),
  "resolvedModelId": zod.string().optional(),
  "visibility": zod.enum(['visible', 'hidden']).describe('\* `visible` - visible\n\* `hidden` - hidden').default(generationDefinitionSchemaStepsByIdVisibilityDefault),
  "groupMode": zod.enum(['none', 'group', 'breakout']).describe('\* `none` - none\n\* `group` - group\n\* `breakout` - breakout').default(generationDefinitionSchemaStepsByIdGroupModeDefault),
  "styleTemplateId": zod.string().nullish(),
  "groupTemplateId": zod.string().nullish(),
  "label": zod.string().nullish(),
  "filter": zod.unknown().nullish()
}))
})

export type GenerationDefinitionSchema = zod.input<typeof GenerationDefinitionSchema>;
export type GenerationDefinitionSchemaOutput = zod.output<typeof GenerationDefinitionSchema>;